[init] Init Project Version to v0.9
