+++ src/main/java/tech/odes/hudi/flink/starter/common/config/HoodieTableConnector.java
@@ -1,0 +1,16 @@
+package tech.odes.hudi.flink.starter.common.config;
+
+import org.apache.flink.configuration.Configuration;
+
+public interface HoodieTableConnector {
+
+    String getBasePath();
+
+    String getTargetPath();
+
+    String getName();
+
+    String getTablePath();
+
+    Configuration getConfiguration();
+}
