--- src/main/resources/transform.yaml
@@ -1,18 +1,0 @@
-source:
-  basePath: hdfs://172.16.2.120:9000/hudi
-  table:
-    - namespace: flink
-      name: t1
-      conf:
-        hoodie.datasource.write.partitionpath.field: partition
-sink:
-  basePath: hdfs://172.16.2.120:9000/hudi
-  table:
-    - namespace: flink
-      name: t2
-      initTableIfNotExists: true
-      conf:
-        hoodie.datasource.write.partitionpath.field: partition
-transform: |-
-  create view tp as select * from t1;
-  insert into t2 select * from t1 where `partition` in ('par2', 'par1', 'par3', 'par4');
